ioworker: Make io_worker module into __init__
